Web Speech API crashes onboundary event handling with reload https://bugs.webkit.org/show_bug.cgi?id=124607
Reviewed by Tim Horton.

If the page goes away, we need to cleanup the Mac platform synthesizer object, because
NSSpeechSynthesizer is retained elsewhere to handle the callbacks (so it doesn't automatically
get torn down).

The layout tests for speech rely on a Mock synthesizer, so there is no good way to test this
Mac platform specific behavior.

* platform/mac/PlatformSpeechSynthesizerMac.mm:
(-[WebSpeechSynthesisWrapper invalidate]):
(WebCore::PlatformSpeechSynthesizer::~PlatformSpeechSynthesizer):



Identifier: 142786@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@159523 268f45cc-cd09-0410-ab3c-d52691b4dbfc