Deduplicate shortish Text node strings during tree construction. <https://webkit.org/b/124855>
Let HTMLConstructionSite keep a hash set of already seen strings over
its lifetime. Use this to deduplicate the strings inside Text nodes
for any string up to 64 characters of length.

This optimization already sort-of existed for whitespace-only Texts,
but those are laundered in the AtomicString table which we definitely
don't want to pollute with every single Text. It might be a good idea
to stop using the AtomicString table for all-whitespace Text too.

3.82 MB progression on HTML5-8266 locally.

Reviewed by Anders Carlsson.


Identifier: 143015@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@159764 268f45cc-cd09-0410-ab3c-d52691b4dbfc