Avoid unnecessary copy-on-write in FillLayer style application. <https://webkit.org/b/124878>
We ended up with a lot of cloned StyleBackgroundData objects on
HTML5-8266. This happened because we always forced a copy-on-write
when applying background-image:inherit / background-image:initial.

This patch adds early returns to both functions. In the "inherit"
case, we bail early if the target style already has the same fill
layer data as its parent style.

In the "initial" case, we optimize for the single-FillLayer case
and add an early return if the relevant value is either unset or
equal to the templatized initial value.

2.46 MB progression on HTML5-8266 locally.

Reviewed by Antti Koivisto.


Identifier: 143033@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@159782 268f45cc-cd09-0410-ab3c-d52691b4dbfc