Crash in JSC::ASTBuilder::Expression JSC::Parser<JSC::Lexer<unsigned char> >::parseUnaryExpression<JSC::ASTBuilder>(JSC::ASTBuilder&) https://bugs.webkit.org/show_bug.cgi?id=124886
Reviewed by Sam Weinig.

Source/JavaScriptCore:

Make sure the error macros propagate an existing error before
trying to create a new error message.  We need to do this as
the parser state may not be safe for any specific error message
if we are already unwinding due to an error.

* parser/Parser.cpp:

LayoutTests:

Add tests

* js/parser-syntax-check-expected.txt:
* js/script-tests/parser-syntax-check.js:

Identifier: 143041@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@159790 268f45cc-cd09-0410-ab3c-d52691b4dbfc