[GStreamer] Invalid command line error when visiting www.chessbase.com https://bugs.webkit.org/show_bug.cgi?id=124715
Reviewed by Philippe Normand.

Source/WebCore:

We were not handling the HTTP errors in the WebKit GStreamer
source and therefore the 404 error page was being 'decoded'. As no
decoder could be found (for obvious reasons), playback failed, but
it should be failing for the source not being found instead of the
decoding problem.

Test: http/tests/media/video-error-does-not-exist.html

* platform/graphics/gstreamer/WebKitWebSourceGStreamer.cpp:
(StreamingClient::handleResponseReceived): Handle HTTP errors in
the source and raise a GStreamer error to the pipeline.

LayoutTests:

Copied and adapted for HTTP from
media/video-error-does-not-exist.html.

* http/tests/media/video-error-does-not-exist-expected.txt: Added.
* http/tests/media/video-error-does-not-exist.html: Added.


Identifier: 143061@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@159810 268f45cc-cd09-0410-ab3c-d52691b4dbfc