test-webkit-scripts should show the failing tests and use an appropriate exit code https://bugs.webkit.org/show_bug.cgi?id=124840
Patch by Jozsef Berta <jberta@inf.u-szeged.hu> on 2013-11-29
Reviewed by Ryosuke Niwa.

A fixme in test-webkit-scripts asked that the script should display success or failiure
and exit with a 0 or 1 value accordingly after all of the tests have completed.

* Scripts/test-webkit-scripts:
(ScriptsTester.run_test_script):
The outcome of the currently run script is returned to the main as a boolean value. A boolean is returned,
because at this point we don't need to pass on more information other than success or failiure.
(ScriptsTester.main):
The return values are now stored for each script and when all tests have completed successfully,
the script indicates success and returns 0. Otherwise it will display the name(s) of the failing script(s) and return 1.

Identifier: 143119@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@159868 268f45cc-cd09-0410-ab3c-d52691b4dbfc