Remove unreachable labels for -webkit-margin-*-collapse properties. <https://webkit.org/b/125057>
The following properties are implemented in DeprecatedStyleBuilder
and should not have case labels in the applyProperty() switch:

    -webkit-margin-before-collapse
    -webkit-margin-top-collapse
    -webkit-margin-after-collapse
    -webkit-margin-bottom-collapse

This seems counter-intuitive, but they are actually *not* like other
directional properties. In this case, before/after are only aliases
for top/bottom, and do not depend on writing-mode or text-direction.
See also r68561, where the aliases were originally added.

Reviewed by Anders Carlsson.


Identifier: 143162@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@159911 268f45cc-cd09-0410-ab3c-d52691b4dbfc