<https://webkit.org/b/125559> Move JavaScriptCore off the legacy WebKit availability macros
The legacy WebKit availability macros are verbose, confusing, and provide no benefit over
using the system availability macros directly. The original vision was that they'd serve
a cross-platform purpose but that never came to be.

Map from WebKit version to OS X version based on the mapping in WebKitAvailability.h.
All iOS versions are specified as 7.0 as that is when the JavaScriptCore C API was made
public.

Part of <rdar://problem/15512304>.

Reviewed by Anders Carlsson.

* API/JSBasePrivate.h:
* API/JSContextRef.h:
* API/JSContextRefPrivate.h:
* API/JSObjectRef.h:
* API/JSValueRef.h:


Identifier: 143641@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@160452 268f45cc-cd09-0410-ab3c-d52691b4dbfc