Incorrect handling of non-ASCII characters in WTF::base64DecodeInternal https://bugs.webkit.org/show_bug.cgi?id=125658
Reviewed by Alexey Proskuryakov.

* wtf/text/Base64.cpp:
(WTF::base64EncodeInternal): Use a reference to an array instead of a pointer
for the table argument type so we check the size of the array at the call site.
(WTF::base64DecodeInternal): Ditto. Also added an bounds check before using the
table so we don't load some random value for non-ASCII characters.

Identifier: 143748@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@160580 268f45cc-cd09-0410-ab3c-d52691b4dbfc